+++ programs/bpf_loader/src/lib.rs
@@ -14,7 +14,12 @@ use crate::{
 };
 use log::{log_enabled, trace, Level::Trace};
 use solana_measure::measure::Measure;
-use solana_program_runtime::instruction_processor::{Executor, InstructionProcessor};
+use solana_program_runtime::{
+    ic_logger_msg, ic_msg,
+    instruction_processor::{Executor, InstructionProcessor},
+    invoke_context::Logger,
+    stable_log,
+};
 use solana_rbpf::{
     aligned_memory::AlignedMemory,
     ebpf::HOST_ALIGN,
@@ -35,12 +40,11 @@ use solana_sdk::{
         reject_deployment_of_unresolved_syscalls, requestable_heap_size,
         stop_verify_mul64_imm_nonzero,
     },
-    ic_logger_msg, ic_msg,
     instruction::{AccountMeta, InstructionError},
     keyed_account::{from_keyed_account, keyed_account_at_index, KeyedAccount},
     loader_instruction::LoaderInstruction,
     loader_upgradeable_instruction::UpgradeableLoaderInstruction,
-    process_instruction::{stable_log, ComputeMeter, InvokeContext, Logger},
+    process_instruction::{ComputeMeter, InvokeContext},
     program_utils::limited_deserialize,
     pubkey::Pubkey,
     rent::Rent,
