+++ programs/config/src/config_processor.rs
@@ -2,9 +2,10 @@
 
 use crate::ConfigKeys;
 use bincode::deserialize;
+use solana_program_runtime::ic_msg;
 use solana_sdk::{
     account::{ReadableAccount, WritableAccount},
-    feature_set, ic_msg,
+    feature_set,
     instruction::InstructionError,
     keyed_account::keyed_account_at_index,
     process_instruction::InvokeContext,
