+++ programs/stake/src/stake_state.rs
@@ -4,12 +4,12 @@
 //! * own mining pools
 
 use {
+    solana_program_runtime::ic_msg,
     solana_sdk::{
         account::{AccountSharedData, ReadableAccount, WritableAccount},
         account_utils::{State, StateMut},
         clock::{Clock, Epoch},
         feature_set::stake_merge_with_unmatched_credits_observed,
-        ic_msg,
         instruction::{checked_add, InstructionError},
         keyed_account::KeyedAccount,
         process_instruction::InvokeContext,
