+++ runtime/src/builtins.rs
@@ -1,8 +1,9 @@
 use crate::system_instruction_processor;
+use solana_program_runtime::stable_log;
 use solana_sdk::{
     feature_set,
     instruction::InstructionError,
-    process_instruction::{stable_log, InvokeContext, ProcessInstructionWithContext},
+    process_instruction::{InvokeContext, ProcessInstructionWithContext},
     pubkey::Pubkey,
     stake, system_program,
 };
