+++ sdk/src/process_instruction.rs
@@ -1,6 +1,5 @@
 #![cfg(feature = "full")]
 
-use itertools::Itertools;
 use solana_sdk::{
     account::AccountSharedData,
     compute_budget::ComputeBudget,
@@ -98,151 +97,10 @@ pub trait InvokeContext {
     fn get_return_data(&self) -> (Pubkey, &[u8]);
 }
 
-/// Convenience macro to log a message with an `Rc<RefCell<dyn Logger>>`
-#[macro_export]
-macro_rules! ic_logger_msg {
-    ($logger:expr, $message:expr) => {
-        if let Ok(logger) = $logger.try_borrow_mut() {
-            if logger.log_enabled() {
-                logger.log($message);
-            }
-        }
-    };
-    ($logger:expr, $fmt:expr, $($arg:tt)*) => {
-        if let Ok(logger) = $logger.try_borrow_mut() {
-            if logger.log_enabled() {
-                logger.log(&format!($fmt, $($arg)*));
-            }
-        }
-    };
-}
-
-/// Convenience macro to log a message with an `InvokeContext`
-#[macro_export]
-macro_rules! ic_msg {
-    ($invoke_context:expr, $message:expr) => {
-        $crate::ic_logger_msg!($invoke_context.get_logger(), $message)
-    };
-    ($invoke_context:expr, $fmt:expr, $($arg:tt)*) => {
-        $crate::ic_logger_msg!($invoke_context.get_logger(), $fmt, $($arg)*)
-    };
-}
-
 /// Compute meter
 pub trait ComputeMeter {
     /// Consume compute units
     fn consume(&mut self, amount: u64) -> Result<(), InstructionError>;
     /// Get the number of remaining compute units
     fn get_remaining(&self) -> u64;
-}
-
-/// Log messages
-pub trait Logger {
-    fn log_enabled(&self) -> bool;
-
-    /// Log a message.
-    ///
-    /// Unless explicitly stated, log messages are not considered stable and may change in the
-    /// future as necessary
-    fn log(&self, message: &str);
-}
-
-///
-/// Stable program log messages
-///
-/// The format of these log messages should not be modified to avoid breaking downstream consumers
-/// of program logging
-///
-pub mod stable_log {
-    use super::*;
-
-    /// Log a program invoke.
-    ///
-    /// The general form is:
-    ///
-    /// ```notrust
-    /// "Program <address> invoke [<depth>]"
-    /// ```
-    pub fn program_invoke(
-        logger: &Rc<RefCell<dyn Logger>>,
-        program_id: &Pubkey,
-        invoke_depth: usize,
-    ) {
-        ic_logger_msg!(logger, "Program {} invoke [{}]", program_id, invoke_depth);
-    }
-
-    /// Log a message from the program itself.
-    ///
-    /// The general form is:
-    ///
-    /// ```notrust
-    /// "Program log: <program-generated output>"
-    /// ```
-    ///
-    /// That is, any program-generated output is guaranteed to be prefixed by "Program log: "
-    pub fn program_log(logger: &Rc<RefCell<dyn Logger>>, message: &str) {
-        ic_logger_msg!(logger, "Program log: {}", message);
-    }
-
-    /// Emit a program data.
-    ///
-    /// The general form is:
-    ///
-    /// ```notrust
-    /// "Program data: <binary-data-in-base64>*"
-    /// ```
-    ///
-    /// That is, any program-generated output is guaranteed to be prefixed by "Program data: "
-    pub fn program_data(logger: &Rc<RefCell<dyn Logger>>, data: &[&[u8]]) {
-        ic_logger_msg!(
-            logger,
-            "Program data: {}",
-            data.iter().map(base64::encode).join(" ")
-        );
-    }
-
-    /// Log return data as from the program itself. This line will not be present if no return
-    /// data was set, or if the return data was set to zero length.
-    ///
-    /// The general form is:
-    ///
-    /// ```notrust
-    /// "Program return: <program-id> <program-generated-data-in-base64>"
-    /// ```
-    ///
-    /// That is, any program-generated output is guaranteed to be prefixed by "Program return: "
-    pub fn program_return(logger: &Rc<RefCell<dyn Logger>>, program_id: &Pubkey, data: &[u8]) {
-        ic_logger_msg!(
-            logger,
-            "Program return: {} {}",
-            program_id,
-            base64::encode(data)
-        );
-    }
-
-    /// Log successful program execution.
-    ///
-    /// The general form is:
-    ///
-    /// ```notrust
-    /// "Program <address> success"
-    /// ```
-    pub fn program_success(logger: &Rc<RefCell<dyn Logger>>, program_id: &Pubkey) {
-        ic_logger_msg!(logger, "Program {} success", program_id);
-    }
-
-    /// Log program execution failure
-    ///
-    /// The general form is:
-    ///
-    /// ```notrust
-    /// "Program <address> failed: <program error details>"
-    /// ```
-    pub fn program_failure(
-        logger: &Rc<RefCell<dyn Logger>>,
-        program_id: &Pubkey,
-        err: &InstructionError,
-    ) {
-        ic_logger_msg!(logger, "Program {} failed: {}", program_id, err);
-    }
 }
